refactor: Decompose corsair-mvp.ts god class into 7 engine modules
Extract 1,416-LOC monolith into focused engine modules under src/engine/:
- recon-engine.ts: Read-only observation (fixture/aws/file modes)
- mark-engine.ts: Drift detection with event emission
- raid-engine.ts: Attack simulation + LaneSerializer
- chart-engine.ts: MITRE→NIST→SOC2 framework mapping
- escape-engine.ts: Scope guards, rollback, RAII cleanup
- event-engine.ts: Event querying and aggregation
- plugin-engine.ts: Plugin discovery and registry
- index.ts: Corsair facade class + barrel re-exports

corsair-mvp.ts becomes a thin backwards-compat shim so all 33 test
files (405 tests) continue working with zero import changes.

Also: move RECON/AWS logic from corsair-agent.ts into ReconEngine,
delete dead src/core/plugin-registry.ts, remove stale bin entry.
